modules/locationsapi_post_db/requirements.txt
kafka-python==2.0.2
SQLAlchemy==1.3.19
psycopg2-binary==2.8.5